requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
boto3==1.34.162
openai==1.51.2
mistralai==0.4.2
ffmpeg-python==0.2.0
pytest==8.3.2